pyproject.toml
[build-system]
requires = [ "setuptools", "setuptools-scm" ]
build-backend = "setuptools.build_meta"


[project]
name = "meeting-summary"
dynamic = [
    "version"
]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Mingyang Bao", email = "mingyangbob@gmail.com"}
]
maintainers = [
    {name = "Mingyang Bao", email = "mingyangbob@gmail.com"}
]
dependencies = [
    "openai-whisper>=20250625",
    "requests>=2.32.5",
    'litellm>=1.79.1'

]
[project.optional-dependencies]
dev = [
    'commitizen',
    "pre-commit",
    "ruff",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-flask",
    "coverage[toml]",
]
web = [
    "flask>=3.1.2",
]

# ===================
# Build
# ===================
[tool.setuptools_scm]
version_file = "src/meeting_summary/__version__.py"

[tool.setuptools.packages.find]
where = ["src"]
include = ["meeting_summary"]

# ===================
# Pytest
# ===================
[tool.pytest.ini_options]
filterwarnings = [
    # "error",
    "ignore::UserWarning",
]
minversion = "6.0"
addopts = """
-v -q
-ra
-s
--cov=src --cov=scripts --cov-report html --no-cov-on-fail --cov-append
--doctest-modules --doctest-continue-on-failure
"""
testpaths = [
    "tests",
]
# ===================
# Coverage
# ===================
[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    "if 0:",
    "if __name__ == '__main__':",
    "@not_implemented",
    "@abstractmethod",
    "pragma: no cover"
]
omit = [
    "__version__.py",# generated by setuptools-scm
    "test_*.py",
    "tests/*",
    "scripts/*",
]
[tool.coverage.html]
directory = "htmlcov"

# ===================
# Ruff
# Refer to https://docs.astral.sh/ruff/rules/
# ===================
[tool.ruff]
target-version = "py313"
line-length = 120
unsafe-fixes = true
exclude = [
    "archived",
    "ignore",
    "input",
    "output",
    ".venv*",
    ".git",
    "**/__version__.py"
]

[tool.ruff.lint]
preview = false
select = ["AIR", "ERA", "YTT", "ANN", "ASYNC", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "FIX", "FA", "INT", "ISC", "ICN", "LOG", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "SLOT", "TID", "TD", "ARG", "PTH", "FLY", "I", "C90", "NPY", "PD", "N", "PERF", "E", "W", "D", "F", "PGH", "PLC", "PLE", "PLR", "PLW", "UP", "FURB", "RUF", "TRY"]
# extend-ignore = ["CPY","DOC","FAST","TC", ]
ignore = [
    "ANN002",#Missing type annotation for
    "ANN003",#Missing type annotation for
    # "ANN101",# Missing type annotation for `cls` in method
    # "ANN102",# Missing type annotation for `cls` in classmethod
    "ANN202",#Missing return type annotation for private function
    "ARG002",# Unused method argument
    "B904",#Within an `except` clause, raise exceptions with `raise ... from err` or `raise ... from None` to distinguish them from errors in exception handling
    "C901",#{name} is too complex ({complexity} > {max_complexity})
    "COM812",
    "D105",#Missing docstring in magic method
    "D205",#1 blank line required between summary line and description
    "D400", # First line should end with a period
    "D415", #missing-terminal-punctuation	First line should end with a period, question mark, or exclamation point
    "D417",#Missing argument description in the docstring
    "DTZ005", #`datetime.datetime.now()` called without a `tz` argument
    "E501",#line too long
    "FBT",#ruff-boolean-trap (FBT)
    "LOG015",#call on root logger
    "N801",# Class name  should use CapWords conventio
    "N802",#Function name  should be lowercase
    "N803",#Argument name should be lowercase
    "N805",#First argument of a method should be named self
    "N815",# Variable  in class scope should not be mixedCase
    "N816", # Variable `use_unit_A_phase_matrix` in global scope should not be mixedCase
    "N818",#Exception name should be named with an Error suffix
    "S101",#Use of assert detected. The enclosed code will be removed when compiling to optimised bytecode.
    "S311",# Standard pseudo-random generators are not suitable for cryptographic purposes
    "S603",#subprocess-without-shell-equals-true (S603)
    "SIM108",#Use ternary operator xxx instead of `if`-`else`-block
    "SLF001",# Private member accessed:
    "RUF012",# Mutable class attributes should be annotated with
    "PLR0912",#Too many branches (15 > 12)
    "PLR0913",#Too many arguments in function definition (7 > 5)
    "PLR0915",#Too many statements (51 > 50)
    "PT012",#`pytest.raises()` block should contain a single simple statement
    "PT018",#Assertion should be broken down into multiple parts
    # incompatible with the formatter
    "Q000",# Single quotes found but double quotes preferred
    "Q003",# Change outer quotes to avoid escaping inner quotes
    "UP017",# Use `datetime.UTC` alias

    # TODO: remove these ignore rules
    "BLE001", # Do not catch blind exception: `Exception`
    "T201", # Use of Print function


    "ANN201",#Missing return type annotation for public function
    "ANN206",#Missing return type annotation for classmethod
    "ANN001",#Missing type annotation for function argument
    "B018",#Found useless expression. Either assign it to a variable or remove it.
    "D203",
    "D100",#Missing docstring in public module
    "D101",#Missing docstring in public class
    "D102",#Missing docstring in public method
    "D103",#Missing docstring in public function
    "D104",#Missing docstring in public package
    "D107",#Missing docstring in __init__
    "D213",# preview
    "D401",#First line should be in imperative mood
    "ERA001",#Found commented-out code
    "N806",# Variable in function should be lowercase
    "PLR2004",# Magic value used in comparison, consider replacing  with a constant variable
    "PLW2901",#loop variable  overwritten by assignment target
    "SIM113",#Use  for index variable
    "TRY300",# Consider moving this statement to an `else` block
    "NPY002",#Replace legacy `np.random.rand` call with `np.random.Generator`
    "INP001",#part of an implicit namespace package. Add an `__init__.py`.
    "TD",# TODO
    "FIX",#FIX
    "EXE",
    ]

[tool.ruff.format]
preview = true
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
